requirements.txt
pydantic>=2.0
openai>=1.0
pyyaml
python-dotenv
